# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Elega API
  version: 0.1.0-draft
  description: 'Elega REST API (/api/v1). This file is the contract (spec-first): packages/api-client is generated
    from it, and an API integration test fails if the server exposes a route not described here. Endpoints are
    implemented milestone by milestone; see docs/progress.md. Conventions: JSON, UTF-8, ISO 8601 UTC timestamps; cursor pagination `{ data, page:
    { nextCursor, hasMore } }` (limit default 20, max 50); unified error format; `Idempotency-Key` on creation endpoints;
    RateLimit-* headers on every response. Realtime events are specified in docs/api/realtime-events.md.'
servers:
- url: https://elega.ru/api/v1
  description: Production
- url: http://localhost:8080/api/v1
  description: Local
security:
- bearer: []
...
paths:
  /auth/register:
    post:
      operationId: postAuthRegister
      tags:
      - auth
      summary: Register (invite code required until open beta)
      description: Creates the account, signs it in (refresh cookie set) and sends a verification
        email. Unverified accounts are read-only until the email is confirmed. Retries are safe
        without an Idempotency-Key because email and username are unique (409 on repeat).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Signed in; refresh token set as the __Host-elega_rt cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/verify-email:
    post:
      operationId: postAuthVerifyEmail
      tags:
      - auth
      summary: Confirm email with one-time token
      description: Accepts both verify_email tokens (sign-up) and change_email tokens (the new
        address becomes the account email). Tokens are single use; an unknown, used or expired
        token returns 400.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/resend-verification:
    post:
      operationId: postAuthResendVerification
      tags:
      - auth
      summary: Resend verification email
      responses:
        '202':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/login:
    post:
      operationId: postAuthLogin
      tags:
      - auth
      summary: Log in with email or username
      description: Wrong identifier and wrong password return the same 401. After repeated
        failures for one identifier the endpoint answers 429 with Retry-After (temporary lockout).
        Suspended and banned accounts get 403 with code account_suspended or account_banned.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/AuthResult'
                - $ref: '#/components/schemas/MfaChallenge'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/login/2fa:
    post:
      operationId: postAuthLogin2fa
      tags:
      - auth
      summary: Complete login with TOTP or recovery code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaLoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/refresh:
    post:
      operationId: postAuthRefresh
      tags:
      - auth
      summary: Rotate refresh token, issue new access token
      parameters:
      - $ref: '#/components/parameters/CsrfHeader'
      security:
      - refreshCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/logout:
    post:
      operationId: postAuthLogout
      tags:
      - auth
      summary: Revoke current session
      parameters:
      - $ref: '#/components/parameters/CsrfHeader'
      security:
      - refreshCookie: []
      responses:
        '204':
          description: Done
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/logout-all:
    post:
      operationId: postAuthLogoutAll
      tags:
      - auth
      summary: Revoke all sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordConfirm'
      responses:
        '204':
          description: Done
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/forgot-password:
    post:
      operationId: postAuthForgotPassword
      tags:
      - auth
      summary: Request reset email (same response whether or not the email exists)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '202':
          description: Done
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/reset-password:
    post:
      operationId: postAuthResetPassword
      tags:
      - auth
      summary: Reset password; revokes all sessions
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Done
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/change-password:
    post:
      operationId: postAuthChangePassword
      tags:
      - auth
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Done
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/change-email:
    post:
      operationId: postAuthChangeEmail
      tags:
      - auth
      summary: Start email change (confirmation to new address)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeEmailRequest'
      responses:
        '202':
          description: Done
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/2fa/setup:
    post:
      operationId: postAuth2faSetup
      tags:
      - auth
      summary: Start TOTP enrolment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordConfirm'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TotpSetup'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/2fa/confirm:
    post:
      operationId: postAuth2faConfirm
      tags:
      - auth
      summary: Confirm TOTP and receive recovery codes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TotpConfirmRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryCodes'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/2fa/disable:
    post:
      operationId: postAuth2faDisable
      tags:
      - auth
      summary: Disable 2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordConfirm'
      responses:
        '204':
          description: Done
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /auth/providers:
    get:
      operationId: getAuthProviders
      tags:
      - auth
      summary: External sign-in providers linked to the account
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LinkedProvider'
                additionalProperties: false
                required:
                - data
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/providers/{provider}:
    delete:
      operationId: deleteAuthProvidersProvider
      tags:
      - auth
      summary: Unlink a provider (refused when it is the only way to sign in)
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/OAuthProvider'
      responses:
        '204':
          description: Done
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/sessions:
    get:
      operationId: getAuthSessions
      tags:
      - auth
      summary: List active sessions
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Page
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
                  page:
                    $ref: '#/components/schemas/PageInfo'
                additionalProperties: false
                required:
                - data
                - page
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/sessions/{id}:
    delete:
      operationId: deleteAuthSessionsId
      tags:
      - auth
      summary: Log out one device (revokes the whole session family)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Done
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT (EdDSA, 15 min)
    refreshCookie:
      type: apiKey
      in: cookie
      name: __Host-elega_rt
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 512
      description: Opaque, HMAC-signed
...
    CsrfHeader:
      name: X-Elega-CSRF
      in: header
      required: true
      schema:
        const: '1'
  responses:
    ValidationFailed:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflicts with existing state (taken email or username, last sign-in method)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found, or hidden from this viewer (blocks and private content return 404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        Retry-After:
          schema:
            type: integer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
              - unauthorized
              - forbidden
              - not_found
              - validation_failed
              - conflict
              - rate_limited
              - payload_too_large
              - unsupported_media_type
              - internal_error
              - mfa_required
              - registration_closed
              - account_suspended
              - account_banned
              - reauth_required
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                additionalProperties: false
                required:
                - message
            requestId:
              type: string
          additionalProperties: false
          required:
          - code
          - message
          - requestId
      additionalProperties: false
      required:
      - error
    PageInfo:
      type: object
      properties:
        nextCursor:
          oneOf:
          - type: string
          - type: 'null'
        hasMore:
          type: boolean
      additionalProperties: false
      required:
      - nextCursor
      - hasMore
    Ok:
      type: object
      properties:
        ok:
          const: true
      additionalProperties: false
      required:
      - ok
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 10
          maxLength: 256
        displayName:
          type: string
          minLength: 1
          maxLength: 64
        username:
          type: string
          pattern: ^[a-z0-9._]{3,30}$
        birthdate:
          type: string
          format: date
        inviteCode:
          type: string
          maxLength: 64
        locale:
          type: string
          enum:
          - ru
          - en
        acceptedTermsVersion:
          type: string
        acceptedPrivacyVersion:
          type: string
        acceptedPdProcessingVersion:
          type: string
          description: Separate consent to personal data processing (152-FZ art. 9)
        captchaToken:
          type: string
      additionalProperties: false
      required:
      - email
      - password
      - displayName
      - username
      - birthdate
      - acceptedTermsVersion
      - acceptedPrivacyVersion
      - acceptedPdProcessingVersion
    LoginRequest:
      type: object
      properties:
        identifier:
          type: string
          description: Email or username
        password:
          type: string
        rememberDevice:
          type: boolean
        captchaToken:
          type: string
      additionalProperties: false
      required:
      - identifier
      - password
    MfaLoginRequest:
      type: object
      properties:
        mfaToken:
          type: string
        code:
          type: string
          pattern: ^([0-9]{6}|[A-Z0-9]{10})$
      additionalProperties: false
      required:
      - mfaToken
      - code
    AuthResult:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/Me'
      additionalProperties: false
      required:
      - accessToken
      - expiresIn
      - user
      description: Refresh token is set as the __Host-elega_rt cookie, never in the body.
    MfaChallenge:
      type: object
      properties:
        mfaRequired:
          const: true
        mfaToken:
          type: string
      additionalProperties: false
      required:
      - mfaRequired
      - mfaToken
    TokenRequest:
      type: object
      properties:
        token:
          type: string
      additionalProperties: false
      required:
      - token
    EmailRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      additionalProperties: false
      required:
      - email
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 10
      additionalProperties: false
      required:
      - token
      - newPassword
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 10
      additionalProperties: false
      required:
      - currentPassword
      - newPassword
    ChangeEmailRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newEmail:
          type: string
          format: email
      additionalProperties: false
      required:
      - currentPassword
      - newEmail
    PasswordConfirm:
      type: object
      properties:
        password:
          type: string
          description: Required when the account has a password. Accounts created through VK ID
            or Yandex ID without a password must instead have signed in within the last 10 minutes
            (otherwise 403 reauth_required).
        code:
          type: string
          description: TOTP or recovery code when 2FA is on
      additionalProperties: false
    TotpSetup:
      type: object
      properties:
        otpauthUrl:
          type: string
        qrSvg:
          type: string
      additionalProperties: false
      required:
      - otpauthUrl
      - qrSvg
    TotpConfirmRequest:
      type: object
      properties:
        code:
          type: string
          pattern: ^[0-9]{6}$
      additionalProperties: false
      required:
      - code
    RecoveryCodes:
      type: object
      properties:
        recoveryCodes:
          type: array
          items:
            type: string
          minItems: 10
          maxItems: 10
      additionalProperties: false
      required:
      - recoveryCodes
    OAuthProvider:
      type: string
      enum:
      - vk
      - yandex
      - google
...
    LinkedProvider:
      type: object
      properties:
        provider:
          $ref: '#/components/schemas/OAuthProvider'
        linkedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - provider
      - linkedAt
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceName:
          type: string
        userAgent:
          type: string
        coarseLocation:
          oneOf:
          - type: string
          - type: 'null'
        lastUsedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        current:
          type: boolean
      additionalProperties: false
      required:
      - id
      - lastUsedAt
      - createdAt
      - current
...
    Me:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum:
          - user
          - moderator
          - analyst
          - admin
        isMinor:
          type: boolean
        locale:
          type: string
          enum:
          - ru
          - en
        timezone:
          type: string
        twoFactorEnabled:
          type: boolean
        hasPassword:
          type: boolean
        profile:
          $ref: '#/components/schemas/Profile'
      additionalProperties: false
      required:
      - id
      - email
      - emailVerified
      - username
      - displayName
      - role
      - isMinor
      - locale
      - timezone
    Profile:
      type: object
      properties:
        bio:
          oneOf:
          - type: string
            maxLength: 500
          - type: 'null'
        city:
          oneOf:
          - type: string
          - type: 'null'
        workplace:
          oneOf:
          - type: string
          - type: 'null'
        education:
          oneOf:
          - type: string
          - type: 'null'
        website:
          oneOf:
          - type: string
            format: uri
          - type: 'null'
        birthday:
          oneOf:
          - type: string
            format: date
          - type: 'null'
        pronouns:
          oneOf:
          - type: string
          - type: 'null'
        relationshipStatus:
          oneOf:
          - type: string
          - type: 'null'
        links:
          type: array
          items:
            type: string
            format: uri
          maxItems: 5
        avatar:
          oneOf:
          - $ref: '#/components/schemas/MediaVariants'
          - type: 'null'
        cover:
          oneOf:
          - $ref: '#/components/schemas/MediaVariants'
          - type: 'null'
        fieldAudience:
          type: object
          additionalProperties:
            type: string
            enum:
            - public
            - friends
            - friends_of_friends
            - only_me
            - custom_list
      additionalProperties: false
...
    MediaVariants:
      type: object
      properties:
        blurhash:
          oneOf:
          - type: string
          - type: 'null'
        width:
          type: integer
        height:
          type: integer
        srcset:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                format: uri
              width:
                type: integer
              format:
                type: string
                enum:
                - avif
                - webp
                - jpeg
            additionalProperties: false
      additionalProperties: false
...